requirements.txt
Flask==0.10.1
Flask-Mail==0.9.0
Flask-MySQL==1.2
Flask-PyMongo==0.3.0
Flask-RESTful==0.2.12
Flask-SQLAlchemy==1.0
Flask-Testing==0.4.1
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.4.0
SQLAlchemy==0.9.6
Werkzeug==0.9.6
aniso8601==0.82
argparse==1.2.1
blinker==1.3
cssselect==0.9.1
itsdangerous==0.24
lxml==3.3.5
nose==1.3.3
pymongo==2.6.3
pytz==2014.4
qrcode==5.0.1
requests==2.3.0
six==1.7.2
twill==1.8.0
wsgiref==0.1.2
factory-boy==2.4.1
